# Doc/openapi.yaml
openapi: 3.0.3
info:
  title: MarketCatalyst (FinApp26) Market Data API
  version: 1.0.0
  description: |
    Canonical read API for every screen in MarketCatalyst. This documents the **data
    contract** the frontend consumes — today implemented as direct Firestore
    reads (via `useCollection()` in `app/iq/hooks/useCollection.ts`) rather than
    a deployed REST gateway, so every path below maps 1:1 to a Firestore
    collection written by a NestJS sync job in `backend/src/sync/`. The shapes
    are copied verbatim from the actual job code (not from the earlier planning
    doc `Doc/05_API_Vendor_Tracker.md`, which has drifted from what is really
    implemented — e.g. `analyst_actions` was planned as a per-event
    upgrade/downgrade feed but is actually a consensus-vote snapshot today).

    **Status legend** (see `x-status` on every path):
    - `live` — job exists, writes real data, screen(s) already read it.
    - `planned` — no backend job exists yet; schema is a proposed shape based
      on the named vendor's typical payload. Nothing reads/writes this today.

    **Source legend** (see `x-primary-source` / `x-alternate-source` on every
    path): the vendor actually wired in `backend/src/vendors/`, plus a faster
    or more complete alternate and why it's faster — see each path's
    description for specifics.

    **Swapping a source never changes this document — and fallback is
    automatic and never silent.** `Company` and `MarketMover` (the two
...

...
  contact:
...
servers:
...
paths:
...
  /companies/{ticker}:
    get:
      tags: [Companies]
      summary: Single company by ticker
      x-status: live
      x-consumed-by:
        - "Same screens/features as GET /companies above — the frontend subscribes to the whole companies collection via useCollection() and merges client-side rather than fetching one ticker at a time, so this single-ticker path documents the equivalent Firestore doc shape, not a distinct call site."
      x-primary-source: >-
        Polygon (migrated 2026-07-12): `/v3/reference/tickers` (name, market cap,
        sector, exchange), daily aggs for price/%change, and
        `/vX/reference/financials` timeframe=ttm for diluted EPS → computed
        P/E = price ÷ EPS.
      x-alternate-source: >-
        FMP `/stable/profile` + `/stable/ratios-ttm` + `/stable/stock-peers`
        (fallback). Still the ONLY source for peers, beta, and dividend yield —
        Polygon has no product for those, so they are null on Polygon-served docs.
      x-fallback-behavior: >-
        Same automatic FMP-then-Polygon fallback as `GET /companies` — see
        that path's x-fallback-behavior for the full mechanics. A 404 here
        means neither source has ever heard of the ticker; a 502 means both
        sources errored (rate limit, outage, etc.) rather than simply
        lacking the ticker.
      operationId: getCompany
      parameters:
        - $ref: "#/components/parameters/Ticker"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  # ────────────────────────────────────────────────────────────────────────
  # TICKERS — live (full US universe: reference metadata + daily price)
  # ────────────────────────────────────────────────────────────────────────
...
  /news:
    get:
      tags: [News]
      summary: Company news headlines
      description: |
        Polygon-primary as of 2026-07-08: adds per-ticker sentiment +
        sentiment_reasoning and free-text keywords Finnhub structurally
        lacks, still polled batch every 30 min (no true push either way).
      x-status: live
      x-consumed-by:
        - "Commentary (app/iq/screens/commentary.tsx) — Live tab (merged ahead of mock commentary items), Premarket/After Hours/My names/Macro tabs (filtered by ET hour or ticker, appended to the corresponding mock arrays), NewsDrawer per-ticker history (shown above the mock buildNewsHistory() narrative section)"
        - "Dashboard (app/iq/screens/dashboard.tsx) — Live Market Feed (falls back to mock feed items when none have synced for the tracked universe yet)"
      x-primary-source: >-
        Polygon `/v2/reference/news`, synced every 30 min during market hours,
        80 tickers per run cycling the full universe, last 2 days lookback.
        Adapter-selected via NEWS_SOURCE (see backend/src/adapters/adapters.module.ts);
        falls back to Finnhub `/company-news` (NEWS_FALLBACK_SOURCE) on failure,
        served with a FALLBACK_USED warning when that happens.
      x-alternate-source: >-
        Benzinga real-time newsfeed with per-firm category tagging
        (earnings/M&A/macro classification built in) and near-instant push
        vs. Polygon/Finnhub's 30-min poll cadence; blocked, no key. Neither
        Polygon nor Finnhub sells this — a Benzinga key would be additive,
        not a straight swap.
      x-write-pattern: >-
        IMMUTABLE / append-only. Doc id is `{ticker}_{articleId}` — a
        published article's headline/summary never change, and the same
        article always maps to the same doc id, so re-polling an overlapping
        lookback window never creates a duplicate. Already correct as-built;
        no change needed here.
      operationId: listNews
      parameters:
        - name: ticker
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/NewsArticle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  # ────────────────────────────────────────────────────────────────────────
  # IPOs — live
  # ────────────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      tags: [Admin]
      summary: All users with their effective (expiry-aware) subscription
      description: |
        Staff accounts (ADMIN_EMAIL) are EXCLUDED from this list and from every
        figure in /admin/revenue — the admin is not a customer, and counting it
        shifts plan mix, ARPU and the churn denominator.
      x-status: live
      x-auth: AdminGuard
      operationId: adminListUsers
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 500, maximum: 2000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/AdminUserRow" }
  /admin/subscriptions:
    get:
      tags: [Admin]
      summary: Payment rows joined to their user
      x-status: live
      x-auth: AdminGuard
      x-note: "Returns [] today — the `payments` collection is empty (Stripe is not integrated)."
      operationId: adminListSubscriptions
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 1000, maximum: 5000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions: { type: array, items: { type: object } }
...
  /dividend-history/{ticker}:
    get:
      tags: [Dividends]
      summary: Full declared dividend history + derived yield, CAGR, streak
      description: |
        Replaces a client-side extrapolation that decayed the current amount by
        a hardcoded 6.5%/yr and derived ex-dates from `sym.charCodeAt(0)`.
        `yieldPct` is null for a non-payer AND for a lapsed payer (no payment in
        the trailing 12 months) — both are correct answers, not gaps.
      x-status: live
      x-collection: dividend_history
      operationId: getDividendHistory
      parameters:
        - name: ticker
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /splits/{ticker}:
    get:
      tags: [Dividends]
      summary: Split history
      description: |
        Bars are refetched with `adjusted=true`, so a split silently REWRITES
        price history. Surfacing splits makes that adjustment legible instead of
        looking like a data error.
      x-status: live
      x-collection: splits
      operationId: getSplits
      parameters:
        - name: ticker
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
components:
  parameters:
    Ticker:
      name: ticker
      in: path
      required: true
      schema: { type: string, example: NVDA }
...
  headers:
    RetryAfter:
      description: Seconds to wait before retrying. Present on every 429 response.
      schema: { type: integer, example: 30 }
  responses:
...
    BadRequest:
      description: Invalid request — a query/path parameter failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { statusCode: 400, error: INVALID_PARAMETER, message: "Query parameter 'from' must be an ISO date (YYYY-MM-DD); got 'yesterday'.", timestamp: "2026-07-05T14:00:00Z" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { statusCode: 404, error: NOT_FOUND, message: "No company document found for ticker 'ZZZZ'.", timestamp: "2026-07-05T14:00:00Z" }
    TooManyRequests:
      description: >-
        This endpoint's upstream vendor quota is exhausted (e.g. FMP's 250
        req/day free tier, Polygon's 5 req/min free tier, SEC EDGAR's 10
        req/sec cap). Never silent — always names which vendor throttled and
        includes `Retry-After`.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { statusCode: 429, error: VENDOR_RATE_LIMITED, message: "FMP daily quota exhausted (250 req/day on this plan) — resets at 00:00 UTC.", source: fmp, retryable: true, timestamp: "2026-07-05T14:00:00Z" }
    UpstreamError:
      description: >-
        The upstream vendor returned an error, malformed data, or timed out.
        If a fallback source is configured for this path (see
        `x-fallback-behavior`) it was already tried and also failed —
        `fallbackAttempted` lists every source tried and why each failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            statusCode: 502
            error: UPSTREAM_VENDOR_ERROR
            message: "All sources failed for company NVDA: fmp (HTTP 500 from FMP profile endpoint, retryable), polygon (no data returned (ticker not found on this source), not retryable)"
            retryable: true
            fallbackAttempted:
              - source: fmp
                error: "HTTP 500 from FMP profile endpoint"
              - source: polygon
                error: "no data returned (ticker not found on this source)"
            timestamp: "2026-07-05T14:00:00Z"
    ServiceUnavailable:
      description: The backend itself (or Firestore) is unreachable — distinct from an upstream vendor failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { statusCode: 503, error: SERVICE_UNAVAILABLE, message: "Firestore write failed after 3 retries — see backend logs for the underlying gRPC error.", retryable: true, timestamp: "2026-07-05T14:00:00Z" }
...
  schemas:
...
    AdminUserRow:
      type: object
      properties:
        uid: { type: string }
        name: { type: string, nullable: true }
        email: { type: string, nullable: true }
        planId: { type: string }
        planName: { type: string }
        status: { type: string }
        storedStatus: { type: string, nullable: true }
        expiredSinceLastWrite: { type: boolean }
        subscriptionStartDate: { type: string, nullable: true }
        subscriptionExpiryDate: { type: string, nullable: true }
        daysRemaining: { type: integer, nullable: true }
        joinedDate: { type: string, nullable: true }
        lastLogin: { type: string, nullable: true }
...
    AdapterWarning:
      type: object
      description: >-
        A non-fatal, structured degradation note attached to a document or
        list response — the mechanism that keeps partial/degraded data from
        being indistinguishable from "vendor genuinely has no data." Mirrors
        `backend/src/adapters/types.ts`'s `AdapterWarning`.
      required: [code, message]
      properties:
        code:
          type: string
          enum: [SUB_REQUEST_FAILED, FIELD_NOT_SUPPORTED, FALLBACK_USED, STALE_DATA]
          description: |
            - SUB_REQUEST_FAILED: one of several parallel vendor calls failed; others succeeded (transient, likely retryable on the next scheduled sync).
            - FIELD_NOT_SUPPORTED: this source structurally never has this field — not a failure, a permanent gap on this vendor.
            - FALLBACK_USED: the primary source failed or returned nothing; this document was served by the configured fallback instead.
            - STALE_DATA: returned data is older than expected (e.g. last successful sync, not this run).
        message: { type: string, description: "Always a complete sentence — never blank." }
        field: { type: string, nullable: true, description: "Comma-separated canonical field name(s) affected, when applicable." }
    ListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          required: [count, asOf]
          properties:
            count: { type: integer }
            asOf: { type: string, format: date-time }
            source: { type: string, description: "Vendor that produced this batch" }
            warnings:
              type: array
              default: []
              description: Batch-level degradation notes (e.g. partial-batch sync failures) — empty array, never omitted, when everything succeeded.
              items: { $ref: "#/components/schemas/AdapterWarning" }
    ErrorResponse:
      type: object
      description: >-
        The single error shape used by every 4xx/5xx response in this API.
        `message` is always populated with a complete, specific sentence —
        never empty, never a generic "An error occurred."
      required: [statusCode, error, message, timestamp]
      properties:
        statusCode: { type: integer }
        error:
          type: string
          description: Stable machine-readable code — safe to switch on in client code.
          enum:
            - INVALID_PARAMETER
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - NOT_FOUND
            - VENDOR_RATE_LIMITED
            - UPSTREAM_VENDOR_ERROR
            - SERVICE_UNAVAILABLE
            - NOT_IMPLEMENTED
        message: { type: string, description: "Human-readable, specific explanation — always non-empty." }
        source: { type: string, nullable: true, description: "Vendor/adapter that failed, when this error originated from a vendor call." }
        retryable: { type: boolean, nullable: true, description: "True if retrying later is likely to succeed (rate limit, timeout, 5xx). False for permanent conditions (not found, bad request)." }
        fallbackAttempted:
          type: array
          nullable: true
          description: Present only on UPSTREAM_VENDOR_ERROR when a fallback source was configured and also failed — every attempt, in order.
          items:
            type: object
            properties:
              source: { type: string }
              error: { type: string }
        timestamp: { type: string, format: date-time }
...
    Company:
      type: object
      description: Firestore `companies/{ticker}` — written by companies.job.ts
      properties:
        ticker: { type: string, example: NVDA }
        name: { type: string, example: NVIDIA Corporation }
        price: { type: number, nullable: true }
        pctChange: { type: number, nullable: true }
        marketCap: { type: number, nullable: true, description: "USD, raw dollars (not billions)" }
        beta: { type: number, nullable: true }
        sector: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        exchange: { type: string, nullable: true }
        week52Range: { type: string, nullable: true }
        volume: { type: number, nullable: true }
        averageVolume: { type: number, nullable: true }
        description: { type: string, nullable: true }
        peRatio: { type: number, nullable: true, description: "null when FMP's ratios-ttm plan restriction rejects this symbol" }
        eps: { type: number, nullable: true }
        dividendYield: { type: number, nullable: true }
        dividendPerShare: { type: number, nullable: true }
        peers:
          type: array
          items: { type: string }
        source: { type: string, nullable: true, example: fmp, description: "Which CompanyProfileAdapter actually served THIS document — 'fmp' or 'polygon'. Differs from the configured COMPANY_PROFILE_SOURCE when a FALLBACK_USED warning is present below." }
        warnings:
          type: array
          default: []
          description: Structured degradation notes (see AdapterWarning) — e.g. a fallback was used, or ratios-ttm failed for this symbol. Empty array, never omitted, when the fetch was fully clean.
          items: { $ref: "#/components/schemas/AdapterWarning" }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time, description: "Added 2026-07-19. Timestamp of the FIRST write of this document; preserved on every subsequent write. A document deleted and later re-synced gets a new value." }
        rsRating:
          type: integer
          nullable: true
          minimum: 1
          maximum: 99
          description: >-
            Written by backend/src/sync/rs-rating.job.ts, NOT companies.job.ts
            — an independent, from-scratch approximation of an IBD-style
            relative-strength score (most-recent-quarter-weighted trailing
            returns, ranked 1-99 within TICKER_UNIVERSE), not the literal
            proprietary IBD formula. Null until enough real ohlcv_bars
            history has accumulated for this ticker.
        rsRatingUpdatedAt: { type: string, format: date-time, nullable: true }
        # Computed indicator/score fields (added 2026-07-12) — written by the
        # no-vendor compute jobs (technical-indicators / tech-rating /
        # fundamentals-growth), NOT companies.job.ts. All null until enough real
        # ohlcv_bars history (or a Polygon financials call) exists.
        rsi14: { type: number, nullable: true, description: "RSI(14) from real ohlcv_bars (technical-indicators.job)." }
        macd: { type: number, nullable: true, description: "MACD(12/26) line (technical-indicators.job)." }
        macdSignal: { type: number, nullable: true, description: "MACD 9-period signal line." }
        macdHistogram: { type: number, nullable: true, description: "MACD line minus signal." }
        rvol: { type: number, nullable: true, description: "Relative volume: latest day's volume ÷ 20-day average." }
        technicalsUpdatedAt: { type: string, format: date-time, nullable: true }
        techRating: { type: integer, nullable: true, minimum: 1, maximum: 99, description: "1-99 composite technical rating (tech-rating.job): momentum 50% + price-vs-50d-MA trend 30% + RSI 20%, percentile-ranked. Approximation, not the proprietary IBD formula." }
        sectorRank: { type: integer, nullable: true, description: "This stock's rank WITHIN its sector by techRating (tech-rating.job)." }
        sectorRankTotal: { type: integer, nullable: true, description: "Number of ranked stocks in the sector." }
        techRatingUpdatedAt: { type: string, format: date-time, nullable: true }
        revenueGrowthYoY: { type: number, nullable: true, description: "YoY revenue growth as a decimal (0.064 = 6.4%), from Polygon annual financials (fundamentals-growth.job)." }
        epsGrowthYoY: { type: number, nullable: true, description: "YoY diluted-EPS growth as a decimal." }
        grossMargin: { type: number, nullable: true, description: "Gross margin as a decimal (0.469 = 46.9%)." }
        fundamentalsFiscalYear: { type: string, nullable: true }
        fundamentalsUpdatedAt: { type: string, format: date-time, nullable: true }
...
    NewsArticle:
      type: object
      description: Firestore `news/{ticker}_{articleId}` — written by news.job.ts
      properties:
        ticker: { type: string }
        headline: { type: string }
        summary: { type: string }
        source: { type: string }
        url: { type: string, format: uri }
        category: { type: string, nullable: true }
        sentiment: { type: string, enum: [positive, negative, neutral], nullable: true, description: "Polygon-only — null when served by the Finnhub fallback (structurally absent there)." }
        sentimentReasoning: { type: string, nullable: true, description: "Polygon-only — null when served by the Finnhub fallback." }
        keywords: { type: array, items: { type: string }, description: "Empty when served by the Finnhub fallback." }
        publishedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time, description: "Added 2026-07-19. Timestamp of the FIRST write of this document; preserved on every subsequent write. A document deleted and later re-synced gets a new value." }
...